fix pagination links browser errors and colors
